+++ src/main/java/finalmission/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package finalmission.global;
+package finalmission.global.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
